Moved listeners into their own directory and renamed messaging to dispatch. This time without somehow changing random namespaces in other solutions.
