+++ Shared/Messaging/Assertions/Interfaces/IAssertListenerFound.cs
@@ -1,4 +1,4 @@
-﻿using IdelPog.Messaging.Messaging;
+﻿using IdelPog.Messaging.Listeners;
 
 namespace IdelPog.Messaging.Assertions
 {
