+++ Shared/Messaging/Listeners/IBufferMessenger.cs
@@ -1,4 +1,4 @@
-﻿namespace IdelPog.Messaging.Messaging
+﻿namespace IdelPog.Messaging.Listeners
 {
     public interface IBufferMessenger
     {
